Exported the function to another file and imported it into the working program
